move test snippets in unit-test folder
